[Refactor]: addViewCount 자료형 수정

+++ src/main/kotlin/com/reditus/novelcia/novel/application/NovelViewWriteBackManager.kt
@@ -55,7 +55,7 @@ class NovelViewWriteBackManager(
 
         writeMapSnapshot.forEach { (novelId, count) ->
             transactional {
-                novelRepository.addViewCount(novelId, count)
+                novelRepository.addViewCount(novelId, count.toLong())
             }
         }
     }
